+++ ThoseGuysServices/ThoseGuys-project/src/main/java/io/thoseguys/project/repositories/MobileInsuranceRepository.java
@@ -2,9 +2,6 @@ package io.thoseguys.project.repositories;
 
 import io.thoseguys.project.domain.MobileInsuranceForm;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
