+++ ThoseGuysServices/ThoseGuys-project/src/main/java/io/thoseguys/project/security/SecurityConfig.java
@@ -69,6 +69,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(SecurityConstants.SIGN_UP_URLS).permitAll()
                 .antMatchers("/insurance/**").permitAll()
                 .antMatchers("/calculatePolicyPrice/**").permitAll()
+                .antMatchers("/admin/**").permitAll()
                 .anyRequest().authenticated();
 
         http.addFilterBefore(jwtAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
